requirements.txt
accelerate==0.26.1
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
filelock==3.17.0
fsspec==2024.12.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.63.0rc1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jsonlines==4.0.0
loguru==0.7.3
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
openai==1.65.4
packaging==24.2
proto-plus==1.26.0
protobuf==4.25.6
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.2
sniffio==1.3.1
sympy==1.13.3
tokenizers==0.15.2
torch==2.2.2
tqdm==4.67.1
transformers==4.37.2
typer==0.9.4
typing_extensions==4.12.2
urllib3==2.3.0
win32_setctime==1.2.0
